backend/requirements.txt
Django==4.1
django-allauth==0.51.0
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.5
pycparser==2.21
requests==2.28.1
wrapt==1.14.1
